Finished cart functionality and fixed bugs

+++ src/app/services/storage.service.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@angular/core';
 
+
 @Injectable({
   providedIn: 'root'
 })
@@ -9,20 +10,29 @@ export class StorageService {
 
 
   set = (key: string, value: string | number) : void => {
+    if (typeof window !== 'undefined' && localStorage) {
     localStorage.setItem(key, JSON.stringify(value));
+    }
   }
 
   get = (key: string) : string | number | null => {
-    const item = localStorage.getItem(key);
+    let item;
+    if (typeof window !== 'undefined' && localStorage) {
+    item = localStorage.getItem(key);
+    }
     return item ? JSON.parse(item) : null;
   }
 
   remove = (key: string) : void => {
+    if (typeof window !== 'undefined' && localStorage) {
     localStorage.removeItem(key);
+    }
   }
 
   clear = () : void => {
+    if (typeof window !== 'undefined' && localStorage) {
     localStorage.clear();
+    }
   }
   
 }
